requirements.txt
cffi==1.16.0
cryptography==42.0.8
honcho==1.1.0
pika==1.3.2
pycparser==2.22
python-dotenv==1.0.1
